% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterworthFilter.R
\name{butterworthFilter}
\alias{butterworthFilter}
\title{Combined Cumulative Interpolation Filter}
\usage{
butterworthFilter(data)
}
\arguments{
\item{data}{dataframe of posixct time (time in seconds) and an instantaneous value (per second i.e. cumecs)}
}
\value{
dataframe with timestamp as posixct a trapezoidal integrated rate and quality code if included.
}
\description{
Derivative of overlapping trapezoidally integrated datasets
}
\details{
Useful for detiding or denoising of data
The overlapping averaging windows allow multiple recalculations of the same data
Which provides an estimate of the uncertainty in applying the filter
}
\examples{

# set.seed(999)
# Generate some random data
df <- StevesCoolRandomTS()

# must be hourly intervals
bwf <- butterworthFilter(data.frame(df$Time, df$Signal+df$Noise))
plot(Signal+Noise ~ Time, data = df, col="grey")
lines(Signal ~ Time, data = df, col="red")
lines(Filtered ~ Date, data = bwf, col = "blue")

}
