% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciDespike.R
\name{cciDespike}
\alias{cciDespike}
\title{cci Despkie}
\usage{
cciDespike(spiky, hoursAvg = 3, stdevs = 2, doPlot = FALSE)
}
\arguments{
\item{spiky}{dataframe of posixct time (time in seconds) and an instantaneous value (per second i.e. cumecs)}

\item{hoursAvg}{number of hours to smooth}

\item{stdevs}{number of standard deviations to include}

\item{doPlot}{do plots}
}
\value{
dataframe with timestamp as posixct, and despiked timeseries data
}
\description{
Cuts timeseries data outside 2 standard deviations from the 3 hour ccinterp filtered data
}
\details{
Useful for detiding or denoising of data
The overlapping averaging windows allow multiple recalculations of the same data
Which provides an estimate of the uncertainty in applying the filter
}
\examples{
data(xy_data)
D <- within(xy_data, {
  x <- as.POSIXct(x * 60 * 60 * 24, origin = "1970-01-01")
})

# add some spikes
D[round(nrow(D)/2),2] <- 60
D[round(nrow(D)/4),2] <- 80

plot(D, col="red", pch=16)

# despike
despikedD <- cciDespike(D)
points(despikedD, col="black", pch=16)
}
