% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccInterp.R
\docType{package}
\name{ccInterp}
\alias{ccInterp-package}
\alias{ccInterp}
\title{Combined Cumulative Interpolation Filter}
\description{
Filters a source timeseries to remove noise/tides/etc
Output data as Hourly.
}
\author{
Stephen Wallace \email{stephen.wallace@des.qld.gov.au}
}
