% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/godinFilter.R
\name{godinFilter}
\alias{godinFilter}
\title{Combined Cumulative Interpolation Filter}
\usage{
godinFilter(x)
}
\arguments{
\item{x}{vector of equally spaced interval data}
}
\value{
dataframe with timestamp as posixct a trapezoidal integrated rate and quality code if included.
}
\description{
Derivative of overlapping trapezoidally integrated datasets
}
\details{
Useful for detiding or denoising of data
The overlapping averaging windows allow multiple recalculations of the same data
Which provides an estimate of the uncertainty in applying the filter
}
\examples{
data <- cumsum(rnorm(1000))
data <- data * c(0, data[-length(data)])
plot(data)
lines(godinFilter(data), col="red")
}
