% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxfun.R
\name{maxfun}
\alias{maxfun}
\title{MaxMin Function}
\usage{
maxfun(t, y, x, option = "max")
}
\arguments{
\item{option}{"max" or "min" to return interval max or interval min.}

\item{refx}{source x input.}

\item{refy}{source y input.}

\item{targetx}{target x interval i.e. hourly/daily.}
}
\value{
vector of quality codes or such
}
\description{
Simply a wrapper for approxfun, returning the maximum value forward
and backwards from a lookup table.
}
\examples{

longts <- seq(0, 500, by = 1)
# create a random timeseries
randts <- runif(length(longts), min = 0, max = 1000)

plot(longts, randts)
lines(longts, randts, col = "grey")

# new interval timeseries
intervalts <- seq(0, 500, by = 20)

intervalMax <- maxfun(longts, randts, intervalts)
f.max <- approxfun(intervalMax, method = "constant")
lines(f.max(longts), col = "red")

intervalMin <- maxfun(longts, randts, intervalts, option = "min")
f.min <- approxfun(intervalMin, method = "constant")
lines(f.min(longts), col = "red")

}
