% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StevesCoolRandomTS.R
\name{StevesCoolRandomTS}
\alias{StevesCoolRandomTS}
\title{Steves Cool Random Timeseries}
\usage{
StevesCoolRandomTS(
  maxFlow = 500 * runif(1),
  maxNoise = 500 * runif(1),
  obs = 10000,
  smoothed = TRUE
)
}
\arguments{
\item{maxFlow}{The upper limit of generated freshwater data}

\item{maxNoise}{The upper limit of generated tidal data}

\item{obs}{Roughly how many hours to simulate}

\item{smoothed}{applies godin filter to simulated flow when TRUE}
}
\value{
dataframe with timestamp as posixct a trapezoidal integrated rate and quality code if included.
}
\description{
Generates a timeseries of random freshwater and tidal discharge
}
\details{
Useful for examples and evaluating tide filters
}
\examples{

par(mfrow=c(3,1))
set.seed(999)
smoothPlot <- StevesCoolRandomTS(obs=1000)
set.seed(999)
rough <- StevesCoolRandomTS(obs=1000, smoothed=FALSE)
 plot(smoothPlot$Time, smoothPlot$Signal)
 lines(rough$Time, rough$Signal, col="red")
 plot(smoothPlot$Time, smoothPlot$Noise)
 plot(smoothPlot$Time, smoothPlot$Signal+smoothPlot$Noise)
 lines(rough$Time, rough$Signal+rough$Noise, col='red')

}
