% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccInterp.R
\docType{data}
\name{xy2_data}
\alias{xy2_data}
\title{Xy2 data}
\format{
An object of class \code{list} of length 2.
}
\usage{
xy2_data
}
\description{
An data set used in the examples
}
\details{
A data set of 457 rows and 2 columns
\describe{
  \item{x}{A numeric vector}
  \item{y}{A numeric vector}
}
}
\keyword{datasets}
