% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccInterp.R
\docType{data}
\name{xy_data}
\alias{xy_data}
\title{Xy data}
\format{
An object of class \code{data.frame} with 457 rows and 2 columns.
}
\usage{
xy_data
}
\description{
An data set used in the examples
}
\details{
A data set of 457 rows and 2 columns
\describe{
  \item{x}{A numeric vector}
  \item{y}{A numeric vector}
}
}
\keyword{datasets}
