% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciDespike.R
\name{cciDespike}
\alias{cciDespike}
\title{cci Despkie}
\usage{
cciDespike(spiky, hoursAvg = 3, stdevs = 2, doPlot = FALSE)
}
\arguments{
\item{spiky}{dataframe of posixct time (time in seconds) and an instantaneous value (per second i.e. cumecs)}

\item{hoursAvg}{number of hours to smooth}

\item{stdevs}{number of standard deviations to include}

\item{doPlot}{do plots}
}
\value{
dataframe with timestamp as posixct, and despiked timeseries data
}
\description{
Cuts timeseries data outside 2 standard deviations from the 3 hour ccinterp filtered data
}
\details{
Worth experimenting with the number of standard deviations and averaging window for the particular dataset
}
\examples{
data(xy_data)
D <- within(xy_data, {
  x <- as.POSIXct(x * 60 * 60 * 24, origin = "1970-01-01")
})

# add some spikes
D[round(nrow(D)/2),2] <- 60
D[round(nrow(D)/4),2] <- 80

plot(D, col="red", pch=16)

# despike
despikedD <- cciDespike(D)
points(despikedD, col="black", pch=16)

spiky <- StevesCoolRandomTS()
spiky <- spiky \%>\% mutate(Tided = Signal+Noise+TideInteraction) \%>\%
dplyr::select(Time, Tided)
dfoutliers <- sample_n(spiky, 20) \%>\%
  mutate( Tided = rnorm(20, 0, 200) )

spiky <- spiky \%>\%
left_join(dfoutliers, by = "Time", suffix = c("", "_new")) \%>\%
mutate(Tided = dplyr::coalesce(Tided_new, Tided)) \%>\%
dplyr::select(Time, Tided)

cciDespike(spiky, doPlot = TRUE, stdevs = 2)
points(dfoutliers, col = "blue", pch = 4)

}
